correct some bug  add searchable.java
