--- app/src/main/java/itac/article/searchresults_activity.java
@@ -1,33 +1,0 @@
-package itac.article;
-
-import android.app.Activity;
-import android.app.SearchManager;
-import android.content.Intent;
-import android.os.Bundle;
-
-/**
- * Created by user on 16/2/20.
- */
-public class searchresults_activity extends Activity {
-
-    @Override
-    public void onCreate(Bundle savedInstanceState) {
-
-        handleIntent(getIntent());
-    }
-
-    @Override
-    protected void onNewIntent(Intent intent) {
-
-        handleIntent(intent);
-    }
-
-    private void handleIntent(Intent intent) {
-
-        if (Intent.ACTION_SEARCH.equals(intent.getAction())) {
-            String query = intent.getStringExtra(SearchManager.QUERY);
-            //use the query to search your data somehow
-        }
-    }
-    
-}
