+++ app/src/main/res/menu/menu_main.xml
@@ -16,5 +16,5 @@
         android:title="@string/search_title"
         android:icon="@drawable/ic_menu_search"
         app:showAsAction="collapseActionView|ifRoom"
-        app:actionViewClass="android.widget.SearchView" />
+        app:actionViewClass="android.support.v7.widget.SearchView" />
 </menu>
