+++ app/src/main/res/values/strings.xml
@@ -16,5 +16,5 @@
     <string name="spam_string">Spams</string>
     <string name="closeDrawer">Navigation View</string>
     <string name="openDrawer">Navigation View Closed</string>
-    <string name="search_title">search title</string>
+    <string name="search_title">search</string>
 </resources>
